refactor(gui): split the decision detail view by responsibility
